Moving revisions action/reducer/selector into its own subdir under `state/posts`
