+++ client/state/posts/actions.js
@@ -23,10 +23,6 @@ import {
 	POST_SAVE,
 	POST_SAVE_SUCCESS,
 	POST_SAVE_FAILURE,
-	POST_REVISIONS_RECEIVE,
-	POST_REVISIONS_REQUEST,
-	POST_REVISIONS_REQUEST_FAILURE,
-	POST_REVISIONS_REQUEST_SUCCESS,
 	POSTS_RECEIVE,
 	POSTS_REQUEST,
 	POSTS_REQUEST_SUCCESS,
@@ -55,19 +51,6 @@ export function receivePosts( posts ) {
 	return {
 		type: POSTS_RECEIVE,
 		posts
-	};
-}
-
-export function receiveRevisions( siteId, postId, revisions ) {
-	// NOTE: We expect all revisions to be on the same post, thus highly
-	// coupling it to how the WP-API returns revisions, instead of being able
-	// to "receive" large (possibly unrelated) batch of revisions.
-
-	return {
-		type: POST_REVISIONS_RECEIVE,
-		siteId,
-		postId,
-		revisions
 	};
 }
 
@@ -147,37 +130,6 @@ export function requestSitePost( siteId, postId ) {
 				error
 			} );
 		} );
-	};
-}
-
-export function requestSitePostRevisions( siteId, postId ) {
-	return ( dispatch ) => {
-		dispatch( {
-			type: POST_REVISIONS_REQUEST,
-			siteId,
-			postId
-		} );
-
-		const query = {
-			apiNamespace: 'wp/v2'
-		};
-
-		return wpcom.req.get( `/sites/${ siteId }/posts/${ postId }/revisions`, query )
-			.then( ( revisions ) => {
-				dispatch( receiveRevisions( siteId, postId, revisions ) );
-				dispatch( {
-					type: POST_REVISIONS_REQUEST_SUCCESS,
-					siteId,
-					postId
-				} );
-			} ).catch( ( error ) => {
-				dispatch( {
-					type: POST_REVISIONS_REQUEST_FAILURE,
-					siteId,
-					postId,
-					error
-				} );
-			} );
 	};
 }
